CSL: continued work on the --with-arithlib version, where the biggest step this time is that I now have a version of the "inorm" function used when Reduce performs big-float arithmetic. Also as I work on all this I take the opportunity to try to tidy up, refresh and bring the code base more up the plausible C++ style. Now my very first attempt to see what all this does for performance is miserable and shows a big slow-down. Well the last time I saw something like that it was down to a pile of debugging code that was there to trap errors, but as I make stuff actually work I am going to need to see if I can get proper performance!
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6421 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/isprime.h
@@ -237,7 +237,7 @@ constexpr inline bool constexpr_isprime(uint64_t n)
 }
 
 constexpr inline auto goodPrime(int n)
-{   uint64_t p = (static_cast<uint64_t>(1)<<n) - 1;
+{   uint64_t p = (1ULL<<n) - 1;
     while (!constexpr_isprime(p)) p -= 2;
     return p;
 }
